Struct safe_dns::dns_operations::DnsOperations

pub struct DnsOperations {
    // some fields omitted
}

This is a representational structure for all safe-dns operations

Methods

impl DnsOperations

fn new(client: Arc<Mutex<Client>>) -> Result<DnsOperations, DnsError>

Create a new instance of DnsOperations. It is intended that only one of this be created as it operates on global data such as files.

fn new_unregistered(unregistered_client: Arc<Mutex<Client>>) -> DnsOperations

Create a new instance of DnsOperations. This is used for an unregistered client and will have very limited set of functionalities - mostly reads. This is ideal for browsers etc., which only want to fetch from the Network, not mutate it. It is intended that only one of this be created as it operates on global data such as files.

fn register_dns(&self, long_name: String, public_messaging_encryption_key: &PublicKey, secret_messaging_encryption_key: &SecretKey, services: &Vec<(String, DirectoryKey)>, owners: Vec<PublicKey>, private_signing_key: &SecretKey, data_encryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<StructuredData, DnsError>

Register one's own Dns - eg., pepsico.com, spandansharma.com, krishnakumar.in etc

fn delete_dns(&self, long_name: &String, private_signing_key: &SecretKey) -> Result<StructuredData, DnsError>

Delete the Dns-Record

fn get_all_registered_names(&self) -> Result<Vec<String>, DnsError>

Get all the Dns-names registered by the user so far in the network.

fn get_messaging_encryption_keys(&self, long_name: &String) -> Result<(PublicKey, SecretKey), DnsError>

Get the messaging encryption keys that the user has associated with one's particular Dns-name.

fn get_all_services(&self, long_name: &String, data_decryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<Vec<String>, DnsError>

Get all the services (www, blog, micro-blog etc) that user has associated with this Dns-name

fn get_service_home_directory_key(&self, long_name: &String, service_name: &String, data_decryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<DirectoryKey, DnsError>

Get the home directory (eg., homepage containing HOME.html, INDEX.html) for the given service.

fn add_service(&self, long_name: &String, new_service: (String, DirectoryKey), private_signing_key: &SecretKey, data_encryption_decryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<StructuredData, DnsError>

Add a new service for the given Dns-name.

fn remove_service(&self, long_name: &String, service_to_remove: String, private_signing_key: &SecretKey, data_encryption_decryption_keys: Option<(&PublicKey, &SecretKey, &Nonce)>) -> Result<StructuredData, DnsError>

Remove a service from the given Dns-name.